refactor: move atomic deployments test into deployment rejection tests
shouldDoAtomicDeployments was located in CreateDeploymentTest but can be better represented in DeploymentRejectionTest. There, I moved the test case into DeploymentRejectionTest.
