+++ engine/src/test/java/io/camunda/zeebe/engine/processing/deployment/DeploymentRejectionTest.java
@@ -9,6 +9,7 @@ package io.camunda.zeebe.engine.processing.deployment;
 
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.tuple;
 
 import io.camunda.zeebe.engine.util.EngineRule;
 import io.camunda.zeebe.model.bpmn.Bpmn;
@@ -18,9 +19,13 @@ import io.camunda.zeebe.protocol.record.ExecuteCommandResponseDecoder;
 import io.camunda.zeebe.protocol.record.Record;
 import io.camunda.zeebe.protocol.record.RecordType;
 import io.camunda.zeebe.protocol.record.RejectionType;
+import io.camunda.zeebe.protocol.record.intent.DeploymentDistributionIntent;
 import io.camunda.zeebe.protocol.record.intent.DeploymentIntent;
+import io.camunda.zeebe.protocol.record.intent.ProcessIntent;
 import io.camunda.zeebe.protocol.record.value.DeploymentRecordValue;
+import io.camunda.zeebe.test.util.record.RecordingExporter;
 import io.camunda.zeebe.test.util.record.RecordingExporterTestWatcher;
+import java.util.stream.Collectors;
 import org.junit.ClassRule;
 import org.junit.Rule;
 import org.junit.Test;
@@ -255,5 +260,39 @@ public class DeploymentRejectionTest {
 
     assertThat(deploymentRejection.getRejectionReason())
         .contains("Must contain at least one executable process");
+  }
+
+  @Test
+  public void shouldDoAtomicDeployments() {
+    // given
+    final String process1Id = "process1";
+    final BpmnModelInstance process1 =
+        Bpmn.createExecutableProcess(process1Id)
+            .startEvent()
+            .businessRuleTask("invalid_business_rule_task_id")
+            .endEvent()
+            .done();
+    final String process2Id = "process2";
+    final BpmnModelInstance process2 =
+        Bpmn.createExecutableProcess(process2Id).startEvent().task().endEvent().done();
+
+    // when
+    ENGINE
+        .deployment()
+        .withXmlResource(process1)
+        .withXmlResource(process2)
+        .expectRejection()
+        .deploy();
+
+    // then
+    assertThat(
+        RecordingExporter.records()
+            .limit(r -> r.getRejectionType() == RejectionType.INVALID_ARGUMENT)
+            .collect(Collectors.toList()))
+        .extracting(Record::getIntent, Record::getRecordType)
+        .doesNotContain(
+            tuple(ProcessIntent.CREATED, RecordType.EVENT),
+            tuple(DeploymentIntent.CREATED, RecordType.EVENT),
+            tuple(DeploymentDistributionIntent.DISTRIBUTING, RecordType.EVENT));
   }
 }
